Move location hierarchy
